Refactor QueueInfo to prepare for wave counters
